Архитектура администратора  и создание utils.cpp
