+++ sources/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(utils)
 add_subdirectory(client)
 add_subdirectory(server)
 add_subdirectory(fileHandler)
