+++ sources/server/CMakeLists.txt
@@ -8,7 +8,7 @@ list(APPEND ${PROJECT_NAME}_SOURCES
         adm_menu/menu.cpp
         ${CMAKE_SOURCE_DIR}/sources/fileHandler/file_handler.cpp
         actions/registration.cpp
-        ../utils.cpp
+        ../utils/utils.cpp
 
 )
