+++ sources/server/actions/registration.cpp
@@ -3,15 +3,14 @@
 
 #include "utils.h"
 
-    // TODO (Viktor): Относительный путь
-
 bool Registration::execute(nlohmann::json &cli_json) {
     auto root_path_users = CreateRootDir("database/users/");
-    auto root_path_requests= CreateRootDir("database/requests/");
-    std::filesystem::path checkUser{root_path_users.string() + std::string(cli_json["args"]["name"]) + ".json"};
-    if (!std::filesystem::exists(checkUser)) {
-        std::filesystem::path newUserRequest{root_path_requests.string() + std::string(cli_json["args"]["name"]) + ".json"};
-        FileDB file_db(newUserRequest);
+    std::filesystem::path check_user{root_path_users.string() + std::string(cli_json["args"]["name"]) + ".json"};
+
+    if (!std::filesystem::exists(check_user)) {
+        auto root_path_requests= CreateRootDir("database/requests/");
+        std::filesystem::path new_user_request{root_path_requests.string() + std::string(cli_json["args"]["name"]) + ".json"};
+        FileDB file_db(new_user_request);
         file_db.write(cli_json);
         return true;
     } else {
