+++ sources/server/adm_menu/CMakeLists.txt
@@ -3,8 +3,11 @@ set(PROJECT_NAME Admin)
 # Реализация сервера
 list(APPEND ${PROJECT_NAME}_SOURCES
         menu.cpp
+        users.cpp
+        turn_on.cpp
+        turn_off.cpp
         requests.cpp
-        ${CMAKE_SOURCE_DIR}/sources/utils.cpp
+        ${CMAKE_SOURCE_DIR}/sources/utils/utils.cpp
         ${CMAKE_SOURCE_DIR}/sources/server/server.cpp
         ${CMAKE_SOURCE_DIR}/sources/server/serv_connection.cpp
         ${CMAKE_SOURCE_DIR}/sources/fileHandler/file_handler.cpp
@@ -16,11 +19,11 @@ list(APPEND ${PROJECT_NAME}_SOURCES)
 
 add_executable(start_admin ${${PROJECT_NAME}_SOURCES})
 
-target_include_directories(start_admin PUBLIC "${CMAKE_SOURCE_DIR}/include/"
-                                            "${CMAKE_SOURCE_DIR}/include/server"
-                                            "${CMAKE_SOURCE_DIR}/lib/JSON/include"
-                                            "${CMAKE_SOURCE_DIR}/include/fileHandler"
-                                            "${CMAKE_SOURCE_DIR}/include/userManager"
-                                            "${CMAKE_SOURCE_DIR}/include/server/adm_menu"
+target_include_directories(start_admin PUBLIC   "${CMAKE_SOURCE_DIR}/include/utils"
+                                                "${CMAKE_SOURCE_DIR}/include/server"
+                                                "${CMAKE_SOURCE_DIR}/lib/JSON/include"
+                                                "${CMAKE_SOURCE_DIR}/include/fileHandler"
+                                                "${CMAKE_SOURCE_DIR}/include/userManager"
+                                                "${CMAKE_SOURCE_DIR}/include/server/adm_menu"
 
 )
