+++ sources/server/adm_menu/turn_off.cpp
@@ -1,0 +1,8 @@
+#include "menu.h"
+
+// TODO: Надо еще подумать над реализацией
+
+// Выключение сервера
+int TurnOff::Choice() {
+    exit(0);
+}
