+++ sources/server/adm_menu/turn_on.cpp
@@ -1,0 +1,18 @@
+#include "menu.h"
+#include "server.h"
+
+
+// Включение сервера
+int TurnOn::Choice() {
+    try {
+        Server server(5000);
+        server.Run();
+    }
+    catch (const std::runtime_error err) {
+        std::cerr << err.what() << std::endl;
+    }
+    catch (...) {
+        std::cerr << "[ERROR] [SERVER] Unknown exception" << std::endl;
+    }
+    return 0;
+}
